remove feed like fixed

+++ models/feeds.Model.js
@@ -44,7 +44,7 @@ const FeedsModel = {
         return database.promise().query(query)
     },
     async DeleteFeedLike(data){
-        let query = `delete from feed_likes where click_user_id = ${data}`
+        let query = `delete from feed_likes where click_user_id = ${data.user_id} and feed_id = ${data.feed_id}`
         return database.promise().query(query)
     },
     async GetFeedLike(data){
